Fixed return value from SaveOfficeVisit to be ov_id so it can be shown in view and ready to add diagnosis data to

+++ HealthInformationSystem/Data/OfficeVisitDiagnosisData.cs
@@ -43,11 +43,8 @@ namespace HealthInformationProgram.Data
             return list;
         }
 
-        public int CreateOfficeVisitDiagnosis(Models.OfficeVisitDiagnosisModel officeVisitDiag)
+        public decimal CreateOfficeVisitDiagnosis(Models.OfficeVisitDiagnosisModel officeVisitDiag)
         {
-
-
-
             var officeVisitDiagRepo = new Data.Repositories.OfficeVisitDiagnosisRepository();
             var dataModel = new Data.Tables.ov_diag();
